Add SerializableSemiring and SerializableFst
