+++ rustfst/src/fst_impls/const_fst/mod.rs
@@ -1,5 +1,4 @@
 pub use self::data_structure::ConstFst;
-pub(crate) use self::data_structure::ConstState;
 
 mod converters;
 mod data_structure;
@@ -7,4 +6,4 @@ mod expanded_fst;
 mod fst;
 mod iterators;
 mod misc;
-mod text_parser;
+mod serializable_fst;
