+++ rustfst/src/fst_impls/vector_fst/mod.rs
@@ -8,5 +8,5 @@ mod fst;
 mod iterators;
 mod misc;
 mod mutable_fst;
+mod serializable_fst;
 mod test;
-mod text_parser;
