+++ rustfst/src/fst_traits/mod.rs
@@ -1,23 +1,19 @@
 #[macro_use]
 mod macros;
 mod allocable_fst;
-mod binary_deserializer;
-mod binary_serializer;
 mod expanded_fst;
 mod final_states_iterator;
 mod fst;
 mod iterators;
 mod mutable_fst;
 mod paths_iterator;
-mod text_parser;
+mod serializable_fst;
 
 pub use self::allocable_fst::AllocableFst;
-pub use self::binary_deserializer::BinaryDeserializer;
-pub use self::binary_serializer::BinarySerializer;
 pub use self::expanded_fst::ExpandedFst;
 pub use self::final_states_iterator::FinalStatesIterator;
 pub use self::fst::{CoreFst, Fst};
 pub use self::iterators::{ArcIterator, FstIterator, FstIteratorMut, StateIterator};
 pub use self::mutable_fst::{MutableArcIterator, MutableFst};
 pub use self::paths_iterator::PathsIterator;
-pub use self::text_parser::TextParser;
+pub use self::serializable_fst::SerializableFst;
