+++ rustfst/src/parsers/bin_fst/mod.rs
@@ -1,5 +1,3 @@
-mod const_fst;
 pub(crate) mod fst_header;
 pub(crate) mod utils_parsing;
 pub(crate) mod utils_serialization;
-mod vector_fst;
