+++ rustfst/src/parsers/mod.rs
@@ -1,4 +1,4 @@
-mod bin_fst;
+pub(crate) mod bin_fst;
 pub mod nom_utils;
 pub mod text_fst;
 pub(crate) mod text_symt;
