+++ rustfst/src/tests_openfst/algorithms/project.rs
@@ -3,17 +3,15 @@ use std::fmt::Display;
 use failure::Fallible;
 
 use crate::algorithms::{project, ProjectType};
-use crate::fst_traits::MutableFst;
-use crate::fst_traits::TextParser;
-use crate::semirings::Semiring;
+use crate::fst_traits::{MutableFst, SerializableFst};
+use crate::semirings::SerializableSemiring;
 use crate::semirings::WeaklyDivisibleSemiring;
-
 use crate::tests_openfst::FstTestData;
 
 pub fn test_project_output<F>(test_data: &FstTestData<F>) -> Fallible<()>
 where
-    F: TextParser + MutableFst + Display,
-    F::W: Semiring<Type = f32> + WeaklyDivisibleSemiring,
+    F: SerializableFst + MutableFst + Display,
+    F::W: SerializableSemiring + WeaklyDivisibleSemiring,
 {
     // Project output
     let mut fst_project_output = test_data.raw.clone();
@@ -33,8 +31,8 @@ where
 
 pub fn test_project_input<F>(test_data: &FstTestData<F>) -> Fallible<()>
 where
-    F: TextParser + MutableFst + Display,
-    F::W: Semiring<Type = f32> + WeaklyDivisibleSemiring,
+    F: SerializableFst + MutableFst + Display,
+    F::W: SerializableSemiring + WeaklyDivisibleSemiring,
 {
     // Project input
     let mut fst_project_input = test_data.raw.clone();
