+++ rustfst/src/tests_openfst/io/const_fst_bin_deserializer.rs
@@ -2,13 +2,13 @@ use failure::Fallible;
 use failure::ResultExt;
 
 use crate::fst_impls::{ConstFst, VectorFst};
-use crate::fst_traits::BinaryDeserializer;
-use crate::semirings::Semiring;
+use crate::fst_traits::SerializableFst;
+use crate::semirings::SerializableSemiring;
 use crate::tests_openfst::FstTestData;
 
 pub fn test_const_fst_bin_deserializer<W>(test_data: &FstTestData<VectorFst<W>>) -> Fallible<()>
 where
-    W: Semiring<Type = f32>,
+    W: SerializableSemiring + 'static,
 {
     let parsed_fst_bin = ConstFst::<W>::read(&test_data.raw_const_bin_path)
         .with_context(|_| format_err!("Failed parsing ConstFst Aligned"))?;
@@ -27,7 +27,7 @@ pub fn test_const_fst_aligned_bin_deserializer<W>(
     test_data: &FstTestData<VectorFst<W>>,
 ) -> Fallible<()>
 where
-    W: Semiring<Type = f32>,
+    W: SerializableSemiring + 'static,
 {
     let parsed_fst_bin = ConstFst::<W>::read(&test_data.raw_const_aligned_bin_path)
         .with_context(|_| format_err!("Failed parsing ConstFst Aligned Bin"))?;
