+++ rustfst/src/tests_openfst/io/const_fst_text_serialization.rs
@@ -3,13 +3,13 @@ use tempfile::tempdir;
 
 use crate::algorithms::fst_convert;
 use crate::fst_impls::{ConstFst, VectorFst};
-use crate::fst_traits::{ExpandedFst, TextParser};
-use crate::semirings::Semiring;
+use crate::fst_traits::{ExpandedFst, SerializableFst};
+use crate::semirings::SerializableSemiring;
 use crate::tests_openfst::FstTestData;
 
 pub fn test_const_fst_text_serialization<W>(test_data: &FstTestData<VectorFst<W>>) -> Fallible<()>
 where
-    W: Semiring<Type = f32>,
+    W: SerializableSemiring + 'static,
 {
     let const_fst_ref: ConstFst<_> = test_data.raw.clone().into();
