+++ rustfst/src/tests_openfst/io/vector_fst_bin_serializer.rs
@@ -1,15 +1,14 @@
 use failure::Fallible;
+use tempfile::tempdir;
 
 use crate::fst_impls::VectorFst;
-use crate::fst_traits::{BinaryDeserializer, BinarySerializer};
-use crate::semirings::Semiring;
-
+use crate::fst_traits::SerializableFst;
+use crate::semirings::SerializableSemiring;
 use crate::tests_openfst::FstTestData;
-use tempfile::tempdir;
 
 pub fn test_vector_fst_bin_serializer<W>(test_data: &FstTestData<VectorFst<W>>) -> Fallible<()>
 where
-    W: Semiring<Type = f32>,
+    W: SerializableSemiring + 'static,
 {
     let dir = tempdir()?;
