+++ rustfst/src/tests_openfst/io/vector_fst_text_serialization.rs
@@ -2,13 +2,13 @@ use failure::Fallible;
 use tempfile::tempdir;
 
 use crate::fst_impls::VectorFst;
-use crate::fst_traits::{ExpandedFst, TextParser};
-use crate::semirings::Semiring;
+use crate::fst_traits::{ExpandedFst, SerializableFst};
+use crate::semirings::SerializableSemiring;
 use crate::tests_openfst::FstTestData;
 
 pub fn test_vector_fst_text_serialization<W>(test_data: &FstTestData<VectorFst<W>>) -> Fallible<()>
 where
-    W: Semiring<Type = f32>,
+    W: SerializableSemiring + 'static,
 {
     let dir = tempdir()?;
